add useAuth hook
This hook serves as a way to access the current user as well
as redirection logic by composing React Router hooks.

The hook throws an error if it is not used within an
AuthContext as it is more or less useless if used without it,
so failing fast would prevent a consumer from incorrectly
using the hook.

Note that the test case for this error behaviour clutters the console.
This is expected as far as I know.