add AuthenticatedRoute component
Although it is possible to manage protected access on a page basis using
the useAuth hook, it can be more convenient not to do so.
AuthenticatedRoute is intended to be used within React-Router's
<Routes> declaration -- one can pass the target element as child of
<AuthenticatedRoute> component in the element prop. As such,
we can get a faster and easier overview of which routes are protected
and which ones are not in the <Routes> component.